% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_count.R
\name{snv_count}
\alias{snv_count}
\title{The snv_count function takes as input the vector of SNVs generated by the snv_extraction function (or
a vector with an equivalent structure) and counts the number of snv_vector. It also provides the percentage
value of that mutation.}
\usage{
snv_table <- snv_count(extracted_snv)
}
\arguments{
\item{snv_vector}{a vector of SNVs like the one returned by snv_extraction}
}
\description{
The snv_count function takes as input the vector of SNVs generated by the snv_extraction function (or
a vector with an equivalent structure) and counts the number of snv_vector. It also provides the percentage
value of that mutation.
}
