% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_extraction.R
\name{snv_extraction}
\alias{snv_extraction}
\title{The snv_extraction function in the snvmut package takes as input a VCF file, a reference genome
and a lenght number. The function returns a vector of SNV mutations that is of leght equal to the
input lenght parameter.}
\usage{
extracted_snv <- snv_extraction(vcf_file, context_length, ref_genome)
}
\arguments{
\item{vcf_file}{a VCF object}

\item{context_length}{a single number which must be an odd integer}

\item{ref_genome}{a BSgenome object}
}
\description{
The snv_extraction function in the snvmut package takes as input a VCF file, a reference genome
and a lenght number. The function returns a vector of SNV mutations that is of leght equal to the
input lenght parameter.
}
\examples{
genome_ref <- BSgenome.Hsapiens.UCSC.hg38
path_to_vcf <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(path_to_vcf, "hg38")[1:150]
extracted_snv <- snv_extraction(vcf, 5, genome_ref)
}
