% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_graphics.R
\name{snv_graphics}
\alias{snv_graphics}
\title{The snv_graphics provides a graphical rapresentation of the data table generated by the function
snv_counts. It plots both a bar plot and a pie chart to quickly visualize the distribution and
proportion of the mutations.}
\usage{
snv_plots <- snv_graphics(snv_table)
}
\arguments{
\item{table_snv_counts}{a data frame like the one returned by snv_count}
}
\description{
The snv_graphics provides a graphical rapresentation of the data table generated by the function
snv_counts. It plots both a bar plot and a pie chart to quickly visualize the distribution and
proportion of the mutations.
}
